# api/berry-buddy-openapi.yaml
openapi: 3.0.3
info:
  title: Berry Buddy API
  version: "0.1.0"
servers:
  - url: http://localhost:5000
...
paths:
...
  /vendors:
    get:
      operationId: vendorsList
      tags: [Vendors]
      summary: List vendors
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Array of vendors
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Vendor' }
    post:
      operationId: vendorsCreate
      tags: [Vendors]
      summary: Create vendor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VendorCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vendor' }
  /vendors/{id}:
    get:
      operationId: vendorsGet
      tags: [Vendors]
      summary: Get vendor by id
      security: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vendor' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: vendorsUpdate
      tags: [Vendors]
      summary: Update vendor
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VendorUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vendor' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: vendorsDelete
      tags: [Vendors]
      summary: Delete vendor
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vendor' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /reviews:
    get:
      operationId: reviewsList
      tags: [Reviews]
      summary: List reviews
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Array of reviews
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Review' }
    post:
      operationId: reviewsCreate
      tags: [Reviews]
      summary: Create review (reported_by = current user)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
  /reviews/{id}:
    patch:
      operationId: reviewsUpdate
      tags: [Reviews]
      summary: Update review (owner-only)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
        '404':
          description: Not found or not owner
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      operationId: reviewsDelete
      tags: [Reviews]
      summary: Delete review (owner-only)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
        '404':
          description: Not found or not owner
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      description: Page size
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
      description: Offset for pagination
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
...
    Vendor:
      type: object
      properties:
        vendor_id: { type: string, format: uuid }
        vendor_name: { type: string }
        vendor_type:
          type: string
          enum: [supermarket, farmers_market, fruit_stand, other]
        address: { type: string, nullable: true }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
        zip_code: { type: string, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        created_at: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time, nullable: true }
    VendorCreate:
      type: object
      required: [vendor_name, vendor_type]
      properties:
        vendor_name: { type: string, minLength: 1, maxLength: 200 }
        vendor_type:
          type: string
          enum: [supermarket, farmers_market, fruit_stand, other]
        address: { type: string }
        city: { type: string }
        state: { type: string }
        zip_code: { type: string }
        latitude: { type: number }
        longitude: { type: number }
    VendorUpdate:
      allOf:
        - $ref: '#/components/schemas/VendorCreate'
      description: All fields optional
...
    Review:
      type: object
      properties:
        review_id: { type: string, format: uuid }
        vendor_id: { type: string, format: uuid }
        berry_id: { type: string, format: uuid }
        reported_by: { type: string, format: uuid }
        rating: { type: integer, minimum: 1, maximum: 5 }
        quality_rating: { type: integer, minimum: 1, maximum: 5, nullable: true }
        freshness_rating: { type: integer, minimum: 1, maximum: 5, nullable: true }
        value_rating: { type: integer, minimum: 1, maximum: 5, nullable: true }
        review_text: { type: string, nullable: true }
        visited_date: { type: string, format: date }
        created_at: { type: string, format: date-time, nullable: true }
    ReviewCreate:
      type: object
      required: [vendor_id, berry_id, rating]
      properties:
        vendor_id: { type: string, format: uuid }
        berry_id: { type: string, format: uuid }
        rating: { type: integer, minimum: 1, maximum: 5 }
        quality_rating: { type: integer, minimum: 1, maximum: 5 }
        freshness_rating: { type: integer, minimum: 1, maximum: 5 }
        value_rating: { type: integer, minimum: 1, maximum: 5 }
        review_text: { type: string }
        visited_date: { type: string, format: date }
    ReviewUpdate:
      allOf:
        - $ref: '#/components/schemas/ReviewCreate'
      description: All fields optional
...
security:
  - bearerAuth: []